Update of p6 and p7
But I'm going to remove these and revert back to the originals - so using BayesX but removing the INLA examples. This is because I cannot use MCMCpack in the final example in Practical 7, so I'm giving up with that package entirely...I will store the MCMCpack code elsewhere, so it's not lost forever.